Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> Hi Cherry,
> 
> For some reason I'm unable to access darwin/arm64 platform so far and It's hard for me to identify the root-cause of the wrong stack trace (https://groups.google.com/d/topic/golang-dev/v1TxCJNemPY/discussion). But the wrong stack trace is more likely related to DWARF of arm64 according to the latest update (https://github.com/golang/go/issues/22716). Besides, in my opinion, most parts of stack trace unwinding shared the same code path between linux and darwin form arm64 and we don't have the wrong stack trace issue in linux/arm64 so the issue in darwin is likely unrelated to FP (frame-pointer).
> 
> I'm also experimenting on putting the FP at the top of frame just as you suggested but find that there are some disadvantages of putting FP at the top of frame rather than at the bottom of frame.
> 1. it requires 1 more instruction and will impact both binary size and performance:
>    FP at the bottom:
>    1. STP.w (FP, LR), -OFFSET(SP)
>    2. MOVD SP, FP
>    FP at the top:
>    1. STP.W (FP, LR), -16(SP)
>    2. MOVD SP, FP
>    3. SUB  $(OFFSET-16), SP
> 
> 2. We need to complicate trace-back logicals (espacially func: gentraceback). Currently, trace-back divides architure into 2 catogries via var: usesLR. If we put FP at the top like 386/amd64 architecture then arm64 is going to be treated as "!usesLR" (because sys.MinFrameSize=0). But arm64 is essentially architecture with LR register and LR value may not be pushed into stack (e.g leaf function) for optimization. So we need to complicate trace-back logicals which is designed to handle 386/amd64 which always push return address automatically into stack and the changes may incurr some tricky bugs.
> 
> I think it's doable for arm64 to put FP at the top of frame just like amd64 has done. But the changes may be big and touch many basic features designed for architecture with LR register. The benefit should be larger on architecture with LR to put FP at the bottom of frame.

Thank you for the experiments! I agree that doing what Clang does, i.e. putting the FP at the top of the frame, doesn't play well with the rest of Go runtime, and make things complicated. This is probably not what we should do.

What about letting FP *pointing* to the top of the current frame, but still storing it at the bottom? Specifically,

  args
  caller's LR
  caller's FP
-------------------   <-- (FP) points here
  locals
  ...
  args for callee
  saved LR
  saved FP
-------------------   <-- (SP) points here

If I understand correctly this is also ABI conforming. This is also consistent with how I (maybe others) understand how frame pointer works.

This can be done with two instructions (for small frames):
STP.W (FP, LR), -framesize(SP)
ADD   $framesize, SP, FP

I think this also plays well with Go's traceback, except the need of telling it LR is saved at 8(SP) instead of 0(SP), which you already did.

Does this mechanism work with Darwin or Linux? In particular, for the case in Issue #22716, I guess the "looping" stack traces are resulted from that the FP is pointing to the current frame, so it thinks the function is called by itself. If FP points to the caller frame, I guess it may help.

Patch-set: 7
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>