Update patch set 3
Patch Set 3: Code-Review-1

(4 comments)

The more I look at this the more I don't want to do it.  It's a very complicated CL to handle a very niche case.

Some of the underlying issue might be fixed by #24416.  If that happens, then remove-the-store-of-what-we-just-loaded optimization should kick in here.  And maybe that would get us most of the way to what you want.

In the meantime, it is easy enough to cover the underlying concern with a test.  For example, you could do:

func TestResetUnwanted(t *testing.T) {
	// Note: unkeyed constructor. It will not compile if new fields are added.
	x := T{1, 1, 1, 1, 1, 1, 1}
	// Reset unexported fields.
	ResetUnwanted(&x)
	// Reset exported fields with reflect (or any other method to enumerate the "wanted" fields).
	v := reflect.ValueOf(&x).Elem()
	for i := 0; i < v.Type().NumField(); i++ {
		if v.Type().Field(i).PkgPath == "" { // exported field
			v.Field(i).Set(reflect.Zero(v.Field(i).Type()))
		}
	}

	// Everything should now be zero.
	if x != (T{}) {
		t.Errorf("ResetUnwanted missed a field %v\n", x)
	}
}

Patch-set: 3
Reviewer: Gerrit User 5200 <5200@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1