Update patch set 37
Patch Set 37:

Gerrit is probably not the best place to develop new code from scratch. 
Uploading 37 patch sets creates a lot of noise for potential reviewers
and it doesn't seem like you are ready for review.

I would encourage you to work on this in a separate, go get'able repo
like maybe a personal GitHub project. It might also help to ask some
other Go developers look at that code or collaborate too, since the
limited discussion here does seem like it has helped.

Separately, it's not a sure thing that we would accept this approach
as a replacement for sort.Stable. The existing code is 200 lines and
fairly straightforward. This code is much larger and much more complex.
If you do get your code working in a way that you're happy with
(and that has good test coverage) then probably the next step would
be to file a proposal discussing the tradeoffs in code complexity
compared against the asymptotic and real improvements of time and space.

Patch-set: 37
CC: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>