Update patch set 5
Patch Set 5:

> Thnkas, I already sent some CLs (for each flag individually, to be able to catch regressions more easily) removing the support_* flags, i.e. CL 106595, CL 106235, CL 106095. I plan to work on the rest as well,

Nice, Thanks Tobias.

> but some (e.g. support_sse41 and support_popcount) are also accessed from > cmd/compile/internal/gc which I don't know yet how to handle. Any help there would be > appreciated.

That is where i struggled too when testing the merging. I guess those can wait another release. Easiest for now might be to just set these from the internal/cpu once after init
until there is a better solution.

> Other support_* flags are used quite early in runtime startup, so we might need to perform initialization of the internal/cpu flags even earlier.

I would recommend to leave the initialization right before it matters for consistency (alginit). If some memequal or memcopy are slow in early runtime i dont think that is a problem. Some architectures need to query the OS for flags so we cant have it earlier than hwcap variables being initialized. And being able to query the env variables will allow for cpu feature flag masking (prototype in golang.org/cl/91737). If we are left with some rarely changed code to check for minimum specs e.g. mmx for 386 that is fine as it wont be used for any support or has variables anywhere else as far as i remember.

Patch-set: 5