Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This CL needs to be updated as there appears to be a merge conflict.

This as is depends on https://go-review.googlesource.com/c/go/+/107138/ . I did not mention it here in the comments because Gerrit shows they are related on this very page (it says "Relation chain") ... I expected this CL would be considered after the one with reversing.

I think the dependency on the other CL is why it says that there is a merge conflict.

Patch-set: 1
Reviewer: Gerrit User 15048 <15048@62eb7196-b449-3ce5-99f1-c037f21e1705>