Update patch set 7
Patch Set 7:

> This PR makes go/build assign .ImportPath and .Root for a testdata package in GOPATH properly

What is your definition of properly?

My definition of properly is that files in testdata directories are completely ignored, as described in "go help test":

    The go tool will ignore a directory named "testdata", making it available
    to hold ancillary data needed by the tests.

and "go help packages":

    Directory and file names that begin with "." or "_" are ignored
    by the go tool, as are directories named "testdata".

There is no such thing as a "testdata package". 

I can't figure out what this CL is doing exactly, but I don't think it is implementing the expected, documented behavior. (I agree that crashing is not expected behavior, of course.)

Patch-set: 7
CC: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>