encoding/xml: fix panic on embedded unexported XMLName
By laws of reflection, unexported fields are unreachable by .Value.
XMLName are allowed at any level of an inner struct but the struct
may not be reachable. If XMLName field is found without name,
it cannot be exported and the value is discarded like other fields.
Comment added to underline how the issue arises.

Another XMLName test was incorrectly set up and is fixed.
Various cases added in a specific test.

Fixes #10538

Change-Id: I65b95011eac8d3cae57836dda2e5f47f08fbda8c