Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Thanks for the advice, Alan. This is taking me a bit longer because I was stuck on named types not working properly. I think there is a bug in go/types now; will get to that first.
> 
> Well, I don't know if there is a bug in go/types or not, but what happens is:
> 
> type A foo.B
> 
> A.Underlying() == foo.B.Underlying()
> 
> I've read the go/types source, and for a long time it has "squashed" underlying named types.
> 
> With that rule, it seems to me like it's impossible for vet to know where the non-named underlying type of a type was declared. If we get to a types.Struct, it's too late - it has no object nor position. And if we get a types.Named, there will only be one - the first one encountered.
> 
> What should we do here? Am I using go/types wrong?

No, that's exactly right.  My example ("type A B") oversimplified the problem I was trying to exhibit as I overlooked that "type A B" does not record the name B, only its underlying type.  A better example would have been:

	type A []B
	type B os.PathError
	var _ = A{{""}}

as, in this case, the underlying type of A is not a struct but a slice of named (B) structs with non-local fields. As written, isLocalType would not descend into B, but I don't think it actually matters because you only call isLocalType for struct literals (such as the inner {""}) but not other composite literals (such as the outer A{{""}}), so isLocalType is good enough as you have written it.

It may yet be clearer written with the single return statement form you proposed.

Sorry for the confusion.

Patch-set: 2
Reviewer: Gerrit User 5195 <5195@62eb7196-b449-3ce5-99f1-c037f21e1705>