runtime: improve performance of grow() in mheap.go
When grow() in mheap.go is called, it sets the span entry for every
page in the newly created span. This is unnecessary as the span is
anyway going to be put into the free list/treap, and only the first
and last span entry need be set for a free span. The newly added code
is similar to how allocSpanLocked() in mheap.go trims the extra pages
and puts it back to the free list/treap.

Fixes #25091

Change-Id: I86ec7cfa07106c8e9648bb3929b49d66adbe6a87