Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> While this is a good observation, I'd rather not. The "only set the head and tail span pointers" optimization that we do when freeing spans has been a source of pain for many experimental optimizations to the memory manager. I'd be completely happy taking *out* that optimization.
> 
> Do you have performance numbers showing that this is a non-trivial improvement? If so, I'd much rather leave this h.setSpans and instead avoid the grow / allocLarge / h.setSpans sequence in allocSpanLocked.

There is considerable speedup observed. But the execution time of grow() in most common scenarios are already in fractions of milliseconds that I am not sure how significant this change is. 

I have captured the time that grow() function takes with the existing code and the with new changes, passing various number of pages as argument.

The columns below are: number of pages passed to grow(), human readable total size corresponding to the # pages passed in, execution time of grow() in milliseconds with existing code, execution time of grow() with the new changes, speedup.

#Pages	Total Size	Old code #c5f0104 (ms)	New code (ms)	Speedup
				
1       8KB     0.041   0.019   2.158
2       16KB    0.076   0.031   2.452
4       32KB    0.078   0.027   2.889
8       64KB    0.088   0.029   3.034
16      128KB   0.062   0.024   2.583
32      256KB   0.08    0.028   2.857
64      512KB   0.086   0.026   3.308
131     ~1MB    0.089   0.03    2.967
640     5MB     0.081   0.027   3.000
1280    10MB    0.084   0.027   3.111
3200    25MB    0.085   0.028   3.036
6400    50MB    0.074   0.032   2.313
12800   100MB   0.141   0.034   4.147
32000   250MB   0.285   0.034   8.382
64000   500MB   0.521   0.049   10.633
131072  1GB     0.955   0.057   16.754
262144  2GB     1.801   0.092   19.576
524288  4GB     3.07    0.13    23.615
1048576 8GB     5.942   0.248   23.960
1179648 9GB     5.822   0.237   24.565
1310720 10GB    6.47    0.253   25.573

Patch-set: 1
Reviewer: Gerrit User 26637 <26637@62eb7196-b449-3ce5-99f1-c037f21e1705>