Update patch set 9
Patch Set 9:

(3 comments)

It is not surprising that a frame is missing due to FP set to caller's frame. I was hoping perf can look at SP if SP is not equal to FP, but it seems not the case. So I guess we have to go with this weird ABI then. Kind of unfortunate.

Seems this doesn't work on darwin/arm64. So probably we only enable frame pointer on linux/arm64? If we add support to other OSes, like *bsd/arm64 or windows/arm64, I don't know what FP usage those platforms expect, and this may or may not work there. My concern is that I think we really don't want to have multiple frame pointer ABI on ARM64.

How much does it affect user assembly code? I'm hoping we don't break users too much. The FP register is already supposed to be reserved, so this is fine. But the frame sizes and offsets need to change. Anything else? Can we provide an automated tool (go fix) that fixes the majority of user assembly code?

Patch-set: 9
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>