Update patch set 9
Patch Set 9:

Another idea (from discussion with Austin): Does it work if we put the saved FP below where SP is pointing to? I.e.

  locals
  ...
  args for callee
  saved LR            <-- (SP) points here
  saved FP            <-- (FP) points here

This way, if Linux perf only looks at FP, not SP, it still sees the chain FPs with all frames. At the same time, 0(SP) is still the saved LR, and all the offsets of locals are unchanged. So we can minimize the affect on assembly code. We may be also able to let the assembler transparently round up the frame size to a multiple of 16, so the frame size in the assembly code doesn't need to change, either. When we grow the stack, we need to reserved the extra word below SP, but the assembler and the runtime should be able to handle it.

Does perf work with this layout?

Patch-set: 9
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>