Update patch set 3
Patch Set 3:

The corresponding unix commit is here: https://go-review.googlesource.com/c/sys/+/109835

This one looks a bit more complicated, because syscall hasn't evolved as much and so is slightly different from unix. However, the additions here are identical to the way it was added in syscall for arm64 and mips64, so this shouldn't be controversial. It's simply "bringing parity."

Patch-set: 3
Reviewer: Gerrit User 20556 <20556@62eb7196-b449-3ce5-99f1-c037f21e1705>