Update patch set 7
Patch Set 7:

> Markdown supports comments. I think that is more appropriate here making it similar to Go copyright headers. And it won't display while rendering the markdown.

My (short) research showed that there isn't a portable way to mark something as a comment. You can only make it an HTML comment, or hide it as a footnote link, or use an extension language's comment (like a Jekyll comment), etc.

Do you have a link or example?

Patch-set: 7