Update patch set 3
Patch Set 3: Code-Review+1

> Patch Set 3: Code-Review-1
> 
> I don't think this is rquired. The runtime and nothing in std uses this so programs should run on Android already. And user programs should not use this package. They should use X/sys/unix.

Isn't that a ticking time bomb? That is - if this apparently unused function is ever used, things will break, and then we'll have to re-figure-it-out. Why not just fix it now?

Also, this has *already* been fixed in syscall for arm64 and mips64. Leaving out amd64 seems silly and is likely to bite us in the future.

Patch-set: 3
Reviewer: Gerrit User 20556 <20556@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+1