Update patch set 3
Patch Set 3:

This does not look like a good use of sync.Pool.  There is no pool here.  I think you are looking for something more like a sharded value, as discussed in https://golang.org/issue/18802.

It's also worth noting that I think a change like this will affect the sequence of random values seen by a program that uses purely sequentially, if the goroutine happens to be scheduled on different threads.  That may break some programs.  It may be OK to make such a change but it shouldn't be done lightly.

Patch-set: 3
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>