Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> This does not look like a good use of sync.Pool.  There is no pool here.  I think you are looking for something more like a sharded value, as discussed in https://golang.org/issue/18802.
> 
> It's also worth noting that I think a change like this will affect the sequence of random values seen by a program that uses purely sequentially, if the goroutine happens to be scheduled on different threads.  That may break some programs.  It may be OK to make such a change but it shouldn't be done lightly.

Sorry, I submitted this prematurely. I've fixed the way the sync pool is used.

I previously discussed this with Filippo - the point of this approach is that it actually *doesn't* change the sequence of random numbers that invokes the global functions purely sequentially. If you take a look at the test output from the latest patch set, you'll see that all the math/rand tests pass, including the existing tests that verify that the output of the RNG is unchanged.


This way, we can avoid mutex contention for projects which invoke the global functions in parallel (oftentimes through imported libraries), but also preserve the expected sequence for programs which invoke it sequentially.

Patch-set: 4
CC: Gerrit User 23704 <23704@62eb7196-b449-3ce5-99f1-c037f21e1705>